refactor: Show grouped search suggestions
- Group search suggestions by media type (Album, Artist, etc.).
- Use a `LazyColumn` to display the suggestions for better performance.
- Update repository methods to fetch items by `Long` ID instead of `String`.
- Handle item clicks on search results.

+++ shared/compose/components/tab-content/src/commonMain/kotlin/com/andannn/melodify/shared/compose/components/tab/content/header/IdBasedGroupHeaderPresenter.kt
@@ -100,9 +100,9 @@ private class GroupHeaderPresenter(
         retainedScope.launch {
             mediaItem =
                 when (val groupKey = groupInfo.groupKey) {
-                    is GroupKey.Artist -> repository.getArtistByArtistId(artistId = groupKey.artistId)
-                    is GroupKey.Album -> repository.getAlbumByAlbumId(albumId = groupKey.albumId)
-                    is GroupKey.Genre -> repository.getGenreByGenreId(genreId = groupKey.genreId)
+                    is GroupKey.Artist -> repository.getArtistByArtistId(artistId = groupKey.artistId.toLong())
+                    is GroupKey.Album -> repository.getAlbumByAlbumId(albumId = groupKey.albumId.toLong())
+                    is GroupKey.Genre -> repository.getGenreByGenreId(genreId = groupKey.genreId.toLong())
                     else -> null
                 }
         }
